feat(license): Add MIT License
The MIT License has been added to the project, which grants permission to use,
copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
software. The justfile has also been added, which contains recipes for
formatting and testing the project. Additionally, a hidden recipe has been added
to the justfile for printing a message in bold square brackets.